fix: widen refresh interval storage from uint8_t to uint16_t
The refresh interval was stored as a single byte, silently truncating any
value above 255. With REFRESH_INTERVAL_MAX = 3600 on M5 Timer Cam-X, user
settings above 255 s were saved incorrectly.

Change EEPROM_ADDR_REFRESH_INTERVAL_LENGTH to 2 bytes and update
SaveRefreshInterval / LoadRefreshInterval to use SaveUint16 / LoadUint16.
Update connect.h/cpp and the WebServer set_int handler to use uint16_t
throughout. LoadRefreshInterval falls back to the factory default when the
stored value is 0 or 0xFFFF (uninitialized EEPROM).

Note: this shifts all subsequent EEPROM addresses by 1 byte — a fresh
flash with Erase All Flash is required.

+++ ESP32_PrusaConnectCam/cfg.h
@@ -32,7 +32,7 @@ public:
   bool CheckActifeWifiCfgFlag();
   void CheckResetCfg();
 
-  void SaveRefreshInterval(uint8_t);
+  void SaveRefreshInterval(uint16_t);
   void SaveToken(String);
   void SaveFingerprint(String);
   void SavePhotoQuality(uint8_t);
@@ -79,7 +79,7 @@ public:
   void SavePrusaLinkIp(String);
   void SavePrusaLinkApiKey(String);
 
-  uint8_t LoadRefreshInterval();
+  uint16_t LoadRefreshInterval();
   String LoadToken();
   String LoadFingerprint();
   uint8_t LoadPhotoQuality();
